requirements.txt
python-dotenv
aiomysql
discord
cryptography